requirements.txt
sentence-transformers
rank-bm25
numpy
torch
pypdf
pytest-cov
